backend/pyproject.toml
[project]
name = "job-application-agent"
version = "0.1.0"
description = "AI-powered job application agent with Gemini + Claude"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # AI / LLM
    "langchain-core>=0.3.0",
    "langchain-google-genai>=4.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-community>=0.3.0",
    # Embeddings & Vector Store
    "chromadb>=0.5.0",
    "langchain-chroma>=0.2.0",
    # Reranking
    "flashrank>=0.2.0",
    # Agent graph
    "langgraph>=0.2.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Config
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    # Task queue
    "arq>=0.26.0",
    # HTTP client
    "httpx>=0.28.0",
    # Multipart form data (file uploads)
    "python-multipart>=0.0.9",
    # PDF generation
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
    # Browser automation
    "browser-use>=0.11.2",
    "playwright>=1.49.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # MCP
    "fastmcp>=3.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-httpx>=0.34.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests requiring real API keys or external services",
    "docker: marks tests requiring Docker",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
